refactor(outbound): recipient form fields (#36899)
