+++ apps/meteor/client/components/Omnichannel/OutboundMessage/components/OutboundMessageWizard/forms/RecipientForm/index.ts
@@ -1,0 +1,2 @@
+export { default } from './RecipientForm';
+export * from './RecipientForm';
